+++ .eslintrc.js
@@ -1,0 +1,17 @@
+require('@rushstack/eslint-patch/modern-module-resolution');
+
+module.exports = {
+  root: true,
+  plugins: ['@cloudflight/typescript'],
+  extends: ['plugin:@cloudflight/typescript/recommended'],
+  ignorePatterns: ['jest.config*.ts'],
+  env: {
+    es6: true,
+    node: true,
+  },
+  parserOptions: {
+    sourceType: 'module',
+    tsconfigRootDir: __dirname,
+    project: ['./packages/*/tsconfig.json', './packages/*/tsconfig.spec.json'],
+  },
+};
