+++ .github/workflows/build.yml
@@ -12,8 +12,10 @@ jobs:
       - uses: actions/checkout@v3
       - uses: volta-cli/action@v4
       - name: Install
-        run: npm ci
+        run: yarn install --immutable --check-cache
       - name: Build
-        run: npm run build
+        run: yarn run build
       - name: Lint
-        run: npm run lint
+        run: yarn run lint
+      - name: Test
+        run: yarn run test
