+++ .github/workflows/publish_to_npm.yml
@@ -8,13 +8,11 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - uses: volta-cli/action@v4
-        with:
-          registry-url: 'https://registry.npmjs.org'
       - name: Install
-        run: npm ci
+        run: yarn install --immutable --check-cache
       - name: Build
-        run: npm run build
+        run: yarn run build
       - name: Publish
-        run: npm run publish-workspaces
+        run: yarn run publish-workspaces
         env:
           NODE_AUTH_TOKEN: ${{ secrets.NPMJS_CLOUDFLIGHT_ADMIN_CI_TOKEN }}
