+++ .yarnrc.yml
@@ -1,0 +1,9 @@
+nodeLinker: node-modules
+
+npmAuthToken: "${NODE_AUTH_TOKEN-''}"
+
+npmRegistryServer: 'https://registry.npmjs.org'
+
+plugins:
+  - path: .yarn/plugins/@yarnpkg/plugin-workspace-tools.cjs
+    spec: '@yarnpkg/plugin-workspace-tools'
