--- packages/stylelint-plugin-angular/tsconfig.lib.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "declaration": true,
-    "module": "CommonJS",
-    "types": []
-  },
-  "include": ["**/*.ts", "**/*.js"],
-  "exclude": ["**/*.spec.ts", "jest.config.js", "jest.config.ts"]
-}
