--- packages/stylelint-plugin-vue/jest.config.ts
@@ -1,16 +1,0 @@
-/* eslint-disable */
-module.exports = {
-  displayName: 'stylelint-plugin',
-  preset: '../../jest.preset.js',
-  globals: {
-    'ts-jest': {
-      tsconfig: '<rootDir>/tsconfig.spec.json',
-    },
-  },
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.[tj]sx?$': 'ts-jest',
-  },
-  moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
-  coverageDirectory: '../../coverage/packages/stylelint-plugin-vue',
-};
