--- packages/stylelint-plugin-vue/project.json
@@ -1,34 +1,0 @@
-{
-  "name": "stylelint-plugin",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "packages/stylelint-plugin-vue/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nrwl/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/packages/stylelint-plugin-vue",
-        "main": "packages/stylelint-plugin-vue/src/index.ts",
-        "tsConfig": "packages/stylelint-plugin-vue/tsconfig.lib.json",
-        "assets": ["packages/stylelint-plugin-vue/*.md"]
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["packages/stylelint-plugin-vue/**/*.ts"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/packages/stylelint-plugin-vue"],
-      "options": {
-        "jestConfig": "packages/stylelint-plugin-vue/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": []
-}
