--- packages/stylelint-plugin-vue/tsconfig.lib.json
@@ -1,11 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "declaration": true,
-    "module": "CommonJS",
-    "types": []
-  },
-  "include": ["**/*.ts", "**/*.js"],
-  "exclude": ["jest.config.ts", "**/*.spec.ts"]
-}
