+++ packages/stylelint-plugin/package.json
@@ -3,6 +3,9 @@
   "version": "0.3.0",
   "author": "Cloudflight GmbH",
   "license": "Apache-2.0",
+  "scripts": {
+    "build-package": "tsc"
+  },
   "keywords": [
     "stylelint",
     "stylelintplugin"
@@ -11,6 +14,10 @@
   "bugs": {
     "url": "https://github.com/cloudflightio/cloudflight-stylelint-plugin/issues"
   },
+  "main": "./dist/index.js",
+  "files": [
+    "dist/**"
+  ],
   "volta": {
     "extends": "../../package.json"
   },
