--- packages/stylelint-plugin/project.json
@@ -1,33 +1,0 @@
-{
-  "name": "stylelint-plugin",
-  "sourceRoot": "packages/stylelint-plugin/src",
-  "projectType": "library",
-  "targets": {
-    "build": {
-      "executor": "@nrwl/js:tsc",
-      "outputs": ["{options.outputPath}"],
-      "options": {
-        "outputPath": "dist/packages/stylelint-plugin",
-        "main": "packages/stylelint-plugin/src/index.js",
-        "tsConfig": "packages/stylelint-plugin/tsconfig.lib.json",
-        "assets": ["packages/stylelint-plugin/*.md"]
-      }
-    },
-    "lint": {
-      "executor": "@nrwl/linter:eslint",
-      "outputs": ["{options.outputFile}"],
-      "options": {
-        "lintFilePatterns": ["packages/stylelint-plugin/**/*.js"]
-      }
-    },
-    "test": {
-      "executor": "@nrwl/jest:jest",
-      "outputs": ["{workspaceRoot}/coverage/packages/stylelint-plugin"],
-      "options": {
-        "jestConfig": "packages/stylelint-plugin/jest.config.ts",
-        "passWithNoTests": true
-      }
-    }
-  },
-  "tags": []
-}
