+++ packages/stylelint-plugin/tsconfig.json
@@ -1,20 +1,29 @@
 {
-  "extends": "../../tsconfig.base.json",
+  "files": ["src/index.js"],
   "compilerOptions": {
-    "allowJs": true,
+    "target": "ES2015",
+    "useDefineForClassFields": true,
+    "module": "CommonJS",
+    "declaration": true,
+    "outDir": "./dist/",
+    "importHelpers": true,
+    "importsNotUsedAsValues": "remove",
+    "stripInternal": true,
+    "isolatedModules": true,
+    "esModuleInterop": true,
     "forceConsistentCasingInFileNames": true,
     "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "exactOptionalPropertyTypes": true,
     "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true
-  },
-  "files": [],
-  "include": [],
-  "references": [
-    {
-      "path": "./tsconfig.lib.json"
-    },
-    {
-      "path": "./tsconfig.spec.json"
-    }
-  ]
+    "noFallthroughCasesInSwitch": true,
+    "noUncheckedIndexedAccess": true,
+    "noImplicitOverride": true,
+    "noPropertyAccessFromIndexSignature": true,
+    "allowUnusedLabels": false,
+    "allowUnreachableCode": false,
+    "skipLibCheck": true,
+    "allowJs": true
+  }
 }
