--- packages/stylelint-plugin/tsconfig.lib.json
@@ -1,15 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "../../dist/out-tsc",
-    "declaration": true,
-    "types": []
-  },
-  "include": ["**/*.ts", "**/*.js"],
-  "exclude": [
-    "**/*.spec.ts",
-    "**/*.spec.js",
-    "jest.config.js",
-    "jest.config.ts"
-  ]
-}
