+++ scripts/prepare.js
