--- tsconfig.base.json
@@ -1,30 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-    "rootDir": ".",
-    "sourceMap": true,
-    "declaration": false,
-    "moduleResolution": "node",
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "importHelpers": true,
-    "target": "es2015",
-    "module": "esnext",
-    "lib": ["es2017", "dom"],
-    "skipLibCheck": true,
-    "skipDefaultLibCheck": true,
-    "baseUrl": ".",
-    "paths": {
-      "@clf-internal/stylelint-plugin-vue": [
-        "packages/stylelint-plugin-vue/src/index.ts"
-      ],
-      "@cloudflight/stylelint-plugin": [
-        "packages/stylelint-plugin/src/index.js"
-      ],
-      "@cloudflight/stylelint-plugin-angular": [
-        "packages/stylelint-plugin-angular/src/index.ts"
-      ]
-    }
-  },
-  "exclude": ["node_modules", "tmp"]
-}
